+++ extensions/youtube/CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## [Updates] - {PR_MERGE_DATE}
 
-- Added optional grid view
+- Added optional grid view. 
 - Fixed title bug by decoding the HTML video and channel titles.
+- Added pinned and recent items for videos and channels.
+- Added metadata to video and channel detail views.